Må opprette egen task for å rydde i oppgaver, ettersom vi ikke har tilgang til oppgavesystemet med våre brukere
